Voter-Commenter Like Issue Resolve

+++ app/controllers/friendship_controller.rb
@@ -20,8 +20,8 @@ class FriendshipController < ApplicationController
 	def friend_request_reject
 	  friendship = Friendship.where(to_id: params[:id], from_id: current_user.id).first
 	  friendship1 = Friendship.where(to_id: current_user.id, from_id: params[:id]).first
-	  unless frienship.nil?
-	  friendship.destroy
+	  unless friendship.nil?
+	  	friendship.destroy
 		end
 	  unless friendship1.nil?
 	  	friendship1.destroy
@@ -38,7 +38,11 @@ class FriendshipController < ApplicationController
 	 	# scope :friendship_id, :conditions=>['to_id = current_user.id OR from_id = current_user.id']
   	# scope :friendship_accept, :conditions=>['accepted = true']
   	# @friends = Friendship.friendship_id.friendship_accept
-	 	 @friends = Friendship.where(to_id: current_user.id, accepted: true)
+
+	 	 friends = Friendship.where(to_id: current_user.id, accepted: true)
+	 	 friends1 = Friendship.where(from_id: current_user.id, accepted: true)
+
+	 	 @fr = friends + friends1 
 	 	# @friends = Friendship.where(from_id: current_user.id OR to_id: current_user.id AND accepted: true)
 		#edirect_to pending_request_path 
 		#@friends = current_user.get_friends
